[pytest] Made the following changes: 1.Use `num_asics()` to get number of asics. 2.Use `sonic-db-cli CONFIG_DB keys 'FEATURE|*' to get the container list. 3.Move two functions to `tests/common/helpers/dut_utils.py`.
