+++ tests/common/helpers/dut_utils.py
@@ -27,3 +27,72 @@ def is_frontend_node(inv_files, hostname):
           node. If we add more types of nodes, then we need to exclude them from this method as well.
     """
     return not is_supervisor_node(inv_files, hostname)
+
+
+def get_group_program_info(duthost, container_name, group_name):
+    """Gets program names, running status and their pids by analyzing the command
+       output of "docker exec <container_name> supervisorctl status". Program name
+       at here represents a program which is part of group <group_name>
+
+    Args:
+        duthost: Hostname of DUT.
+        container_name: A string shows container name.
+        program_name: A string shows process name.
+
+    Returns:
+        A dictionary where keys are the program names and values are their running
+        status and pids.
+    """
+    group_program_info = defaultdict(list)
+    program_name = None
+    program_status = None
+    program_pid = -1
+
+    program_list = duthost.shell("docker exec {} supervisorctl status".format(container_name), module_ignore_errors=True)
+    for program_info in program_list["stdout_lines"]:
+        if program_info.find(group_name) != -1:
+            program_name = program_info.split()[0].split(':')[1].strip()
+            program_status = program_info.split()[1].strip()
+            if program_status in ["EXITED", "STOPPED", "STARTING"]:
+                program_pid = -1
+            else:
+                program_pid = int(program_info.split()[3].strip(','))
+
+            group_program_info[program_name].append(program_status)
+            group_program_info[program_name].append(program_pid)
+
+            if program_pid != -1:
+                logger.info("Found program '{}' in the '{}' state with pid {}"
+                            .format(program_name, program_status, program_pid))
+
+    return group_program_info
+
+
+def get_program_info(duthost, container_name, program_name):
+    """Gets program running status and its pid by analyzing the command
+       output of "docker exec <container_name> supervisorctl status"
+
+    Args:
+        duthost: Hostname of DUT.
+        container_name: A string shows container name.
+        program_name: A string shows process name.
+
+    Return:
+        Program running status and its pid.
+    """
+    program_status = None
+    program_pid = -1
+
+    program_list = duthost.shell("docker exec {} supervisorctl status".format(container_name), module_ignore_errors=True)
+    for program_info in program_list["stdout_lines"]:
+        if program_info.find(program_name) != -1:
+            program_status = program_info.split()[1].strip()
+            if program_status == "RUNNING":
+                program_pid = int(program_info.split()[3].strip(','))
+            break
+
+    if program_pid != -1:
+        logger.info("Found program '{}' in the '{}' state with pid {}"
+                    .format(program_name, program_status, program_pid))
+
+    return program_status, program_pid
